Display the columns correctly under UTF-8
Display the columns correctly in the index screen when the names are
encoded in UTF-8 characters.  This hack is obtained by changing the
way the set_menu_print_format function sets the menu_print_format
variable.  Note also that the configure.ac script must be changed in
order to have the right CFLAGS and LIBS variables being passed to
Makefile.

This patch fixes the bug reported in:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=514933